Create power.java

+++ power.java
@@ -8,6 +8,6 @@ public static void main(String args[])
 Scanner s=new Scanner(System.in);
 int a=s.nextInt();
 int b=s.nextInt();
-int c=Math.pow(a,b);
+int c=(int)Math.pow(a,b);
 System.out.println(c);
 }
